+++ frontend/vue3-web-app/src/components/Fantasy/UserBalance.vue
@@ -6,21 +6,14 @@
 </template>
 
 <script setup lang="ts">
-import { ref, onMounted } from "vue";
+import { computed } from "vue";
 import { useAuthStore } from "@/stores/auth";
-import { localApiService } from "@/services/localApiService";
 import ShardSpan from "../Dom/ShardSpan.vue";
 
 const authStore = useAuthStore();
 
-const userBalance = ref(0);
-
-onMounted(() => {
-    if (authStore.authenticated) {
-        localApiService.getUserBalance().then((result: any) => {
-            userBalance.value = result
-        });
-    }
+const userBalance = computed(() => {
+    return authStore.currentUser.stashBalance ?? 0
 });
 
 </script>
