+++ webapi/csharp-ef-webapi/Controllers/LeagueController.cs
@@ -218,27 +218,5 @@ namespace csharp_ef_webapi.Controllers
 
             return Ok(paginatedMatches);
         }
-
-        // GET: api/league/5/drafts/points
-        [HttpGet("{leagueId}/drafts/points")]
-        [AuthenticatedETag]
-        public async Task<ActionResult<IEnumerable<FantasyDraftPointTotals>>> GetUserDraftFantasyPoints(int leagueId)
-        {
-            try
-            {
-                if (HttpContext.User == null)
-                {
-                    return Ok(new List<FantasyDraftPointTotals>());
-                }
-
-                IEnumerable<FantasyDraftPointTotals> fantasyDraftPointTotals = await _fantasyService.GetFantasyDraftPointTotals(HttpContext.User, leagueId);
-
-                return Ok(fantasyDraftPointTotals);
-            }
-            catch (ArgumentException ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
     }
 }
